change order

+++ football_functions/DTOs/PlayerDTO.cs
@@ -1,4 +1,4 @@
 ﻿using System;
 namespace football_functions.DTOs;
 
-public readonly record struct PlayerDTO(string Name, string Id, double Score);
+public readonly record struct PlayerDTO(string Name, string Id, decimal Score);
